Finalizing refactoring of types into their own files, directories and respective namespaces
