+++ Runtime/Scripts/AbstractInterfaces/IShape.cs
@@ -1,0 +1,9 @@
+﻿using Types;
+
+namespace AbstractInterfaces
+{
+    public interface IShape
+    {
+        Point4 GetPoint4(int i, float resolution, float invResolution);
+    }
+}
