+++ Runtime/Scripts/AbstractInterfaces/IShape.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 41dc66dc637e44b9b854ba66202381b0
+timeCreated: 1658776019
