+++ Runtime/Scripts/AbstractInterfaces/Visualization.cs
@@ -1,4 +1,4 @@
-using Types;
+using Types.Shapes;
 using Unity.Collections;
 using Unity.Jobs;
 using Unity.Mathematics;
@@ -9,51 +9,31 @@ namespace AbstractInterfaces
 {
 	public abstract class Visualization : MonoBehaviour
 	{
-
-		public enum Shape
+		private static readonly Hash.ScheduleShapeDelegate[] ShapeJobs =
 		{
-			Plane
-		  , Sphere
-		  , Torus
-		}
-
-		private static readonly Shapes.ScheduleDelegate[] ShapeJobs =
-		{
-			Shapes.Job<Shapes.Plane>.ScheduleParallel
-		  , Shapes.Job<Shapes.Sphere>.ScheduleParallel
-		  , Shapes.Job<Shapes.Torus>.ScheduleParallel
+			ShapeJob<Types.Shapes.Plane>.ScheduleParallel
+		  , ShapeJob<Types.Shapes.Sphere>.ScheduleParallel
+		  , ShapeJob<Types.Shapes.Torus>.ScheduleParallel
 		};
 
-		private static readonly int
-			PositionsId = Shader.PropertyToID("_Positions");
+		private static readonly int PositionsId = Shader.PropertyToID("_Positions");
+		private static readonly int NormalsId   = Shader.PropertyToID("_Normals");
+		private static readonly int ConfigId    = Shader.PropertyToID("_Config");
 
-		private static readonly int
-			NormalsId   = Shader.PropertyToID("_Normals");
+		[SerializeField]                     private Mesh      instanceMesh;
+		[SerializeField]                     private Material  material;
+		[SerializeField]                     private ShapeType m_shapeType;
+		[SerializeField, Range(0.1f,  10f)]  private float     instanceScale = 2f;
+		[SerializeField, Range(1,     512)]  private int       resolution    = 16;
+		[SerializeField, Range(-0.5f, 0.5f)] private float     displacement  = 0.1f;
 
-		private static readonly int
-			ConfigId    = Shader.PropertyToID("_Config");
-
-		[SerializeField] private Mesh instanceMesh;
-
-		[SerializeField] private Material material;
-
-		[SerializeField] private Shape shape;
-
-		[SerializeField, Range(0.1f, 10f)] private float instanceScale = 2f;
-
-		[SerializeField, Range(1, 512)] private int resolution = 16;
-
-		[SerializeField, Range(-0.5f, 0.5f)] private float displacement = 0.1f;
-
-		private NativeArray<float3x4> m_positions, m_normals;
-
-		private ComputeBuffer m_positionsBuffer, m_normalsBuffer;
-
+		private NativeArray<float3x4> m_positions;
+		private NativeArray<float3x4> m_normals;
+		private ComputeBuffer         m_positionsBuffer;
+		private ComputeBuffer         m_normalsBuffer;
 		private MaterialPropertyBlock m_propertyBlock;
-
-		private bool m_isDirty;
-
-		private Bounds m_bounds;
+		private bool                  m_isDirty;
+		private Bounds                m_bounds;
 
 		private void OnEnable()
 		{
@@ -101,7 +81,7 @@ namespace AbstractInterfaces
 				transform.hasChanged = false;
 
 				UpdateVisualization(m_positions, resolution,
-									ShapeJobs[(int) shape](m_positions, m_normals, resolution, transform.localToWorldMatrix, default));
+									ShapeJobs[(int) m_shapeType](m_positions, m_normals, resolution, transform.localToWorldMatrix, default));
 
 				m_positionsBuffer.SetData(m_positions.Reinterpret<float3>(3 * 4 * 4));
 				m_normalsBuffer.SetData(m_normals.Reinterpret<float3>(3     * 4 * 4));
@@ -113,10 +93,12 @@ namespace AbstractInterfaces
 			Graphics.DrawMeshInstancedProcedural(instanceMesh, 0, material, m_bounds, resolution * resolution, m_propertyBlock);
 		}
 
+		#region Abstract methods
+
 		protected abstract void EnableVisualization(int dataLength, MaterialPropertyBlock propertyBlock);
-
 		protected abstract void DisableVisualization();
-
 		protected abstract void UpdateVisualization(NativeArray<float3x4> positions, int resolution, JobHandle handle);
+
+		#endregion
 	}
 }
