+++ Runtime/Scripts/Types/NoiseType.cs
@@ -1,0 +1,20 @@
+﻿namespace Types
+{
+    public enum NoiseType
+    {
+        Perlin
+      , PerlinTurbulence
+      , Value
+      , ValueTurbulence
+      , Simplex
+      , SimplexTurbulence
+      , SimplexValue
+      , SimplexValueTurbulence
+      , VoronoiWorleyF1
+      , VoronoiWorleyF2
+      , VoronoiWorleyF2MinusF1
+      , VoronoiChebyshevF1
+      , VoronoiChebyshevF2
+      , VoronoiChebyshevF2MinusF1
+    }
+}
