+++ Runtime/Scripts/Types/Point4.cs
@@ -1,0 +1,10 @@
+using Unity.Mathematics;
+
+namespace Types
+{
+	public struct Point4
+	{
+		public float4x3 positions;
+		public float4x3 normals;
+	}
+}
