+++ Runtime/Scripts/Types/Shapes/ShapeType.cs
@@ -1,0 +1,9 @@
+﻿namespace Types.Shapes
+{
+    public enum ShapeType
+    {
+        Plane
+      , Sphere
+      , Torus
+    }
+}
