+++ Runtime/Scripts/Utilities/NoiseUtils.cs
@@ -1,4 +1,7 @@
 using Hash;
+using Types;
+using Unity.Collections;
+using Unity.Jobs;
 using Unity.Mathematics;
 using static Unity.Mathematics.math;
 
@@ -62,6 +65,16 @@ namespace Utilities
             g.c0 += select(-offset, offset, g.c0 < 0f);
             g.c1 += select(-offset, offset, g.c1 < 0f);
             return g;
+        }
+        
+        public static float4x2 IndexTo4UV(int i, float resolution, float invResolution)
+        {
+            float4x2 uv;
+            float4   i4 = 4f * i + float4(0f, 1f, 2f, 3f);
+            uv.c1 = floor(invResolution * i4 + 0.00001f);
+            uv.c0 = invResolution * (i4      - resolution * uv.c1 + 0.5f);
+            uv.c1 = invResolution * (uv.c1                        + 0.5f);
+            return uv;
         }
     }
 }
